feat(9.16): cross-cut full tree base (CatalogTree + refactored Arbols)
