+++ Pages/Documentacion/Arbol.php
@@ -1,14 +1,15 @@
 <?php
 namespace Tuqan\Pages\Documentacion;
 
-use Tuqan\Pages\Catalog\CatalogListado;
+use Tuqan\Pages\Catalog\CatalogTree;
 
 /**
  * Modern tree/arbol view shell for Documentación (Stage 9.12).
  * Grouped view (by tipo_documento/area) as first tree-like experience.
- * Reuses 9.8 helpers. Defers full arbol_documentos logic, editor, perfiles, workflows.
+ * Now uses full tree base (Stage 9.16).
+ * Defers full arbol_documentos logic, editor, perfiles, workflows.
  */
-class Arbol extends CatalogListado
+class Arbol extends CatalogTree
 {
     protected string $table       = 'documentos';
     protected string $title       = 'Árbol de Documentos';
@@ -34,6 +35,9 @@ class Arbol extends CatalogListado
         ];
     }
 
+    /**
+     * Implement for CatalogTree base (Stage 9.16).
+     */
     protected function fetchTreeItems(): array
     {
         $db = $this->getDb();
@@ -47,31 +51,17 @@ class Arbol extends CatalogListado
         return $items;
     }
 
-    protected function buildTreeVariables(array $items): array
+    /**
+     * Implement for CatalogTree base (Stage 9.16).
+     */
+    protected function buildTreeSpecificVariables(array $items): array
     {
-        $base = $this->buildCommonVariables();  // cross-cut helper (Stage 9.13)
-
         // Simple grouping for tree feel (by tipo_documento as top level)
         $grouped = $this->groupItems($items, 'tipo_documento');  // cross-cut helper (Stage 9.13)
 
-        return array_merge($base, [
+        return [
             'documentos' => $items,
             'grouped'    => $grouped,
-        ]);
-    }
-
-    public function ShowPage()
-    {
-        $twig = $this->initTwig();  // cross-cut helper (Stage 9.13)
-
-        $items = $this->fetchTreeItems();
-        $variables = $this->buildTreeVariables($items);
-
-        try {
-            $template = $twig->load($this->templateDir . '/arbol.twig');
-            return $template->render($variables);
-        } catch (\Exception $e) {
-            return "Error al cargar {$this->title}: " . $e->getMessage();
-        }
+        ];
     }
 }
